+++ webpack.config.js
@@ -11,6 +11,7 @@ module.exports = {
       alias: {
         '../../package.json': path.resolve(__dirname, 'package.json'),
         '../../protos/operations.json': path.resolve(__dirname, 'protos/operations.json'),
+        '../../protos/status.json': path.resolve(__dirname, 'protos/status.json'),
       },
     },
     module: {